// src/app/models/system/comaddlocation.ts
export class ComaddlocationModel {
    constructor() {

    }

    company_code: string = "";
    comlocation_code: string = "";
    comaddress_type: string = "";

    comaddlocationth_no: string = "";
    comaddlocationth_moo: string = "";
    comaddlocationth_soi: string = "";
    comaddlocationth_road: string = "";
    comaddlocationth_tambon: string = "";
    comaddlocationth_amphur: string = "";
    provinceth_code: string = "";

    comaddlocationen_no: string = "";
    comaddlocationen_moo: string = "";
    comaddlocationen_soi: string = "";
    comaddlocationen_road: string = "";
    comaddlocationen_tambon: string = "";
    comaddlocationen_amphur: string = "";
    comaddlocation_zipcode: string = "";
    provinceen_code: string = "";

    comaddlocation_tel: string = "";
    comaddlocation_email: string = "";
    comaddlocation_line: string = "";
    comaddlocation_facebook : string = "";

    modified_by: string = "";
    modified_date: string = "";

    index: number = 0;
    select: boolean = false;
    address_ref: string = "";
}